🚚 move [#16] : 폴더 구조 변경
- 폴더 구조 변경
- 폴더 유지를 위한 .gitkeep 파일 추가
