+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/HubRouteDetailResponseDto.java
@@ -1,12 +1,14 @@
-package com.sparta.logistics.hub_service.dto;
+package com.sparta.logistics.hub_service.application.dto;
 
 import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 @Builder
-public class UpdateHubRouteRequestDto {
+@Setter
+public class HubRouteDetailResponseDto {
 
   private UUID startHubId;
   private UUID endHubId;
@@ -14,4 +16,7 @@ public class UpdateHubRouteRequestDto {
   private String endHubName;
   private Long deliveryTime;
   private Long deliveryDistance;
+
+  private Long id; // 임시
+
 }
