+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/UpdateHubRouteRequestDto.java
@@ -1,13 +1,15 @@
-package com.sparta.logistics.hub_service.dto;
+package com.sparta.logistics.hub_service.application.dto;
 
+import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
 @Builder
-public class HubRouteListResponseDto {
+public class UpdateHubRouteRequestDto {
+
+  private UUID startHubId;
+  private UUID endHubId;
   private String startHubName;
   private String endHubName;
   private Long deliveryTime;
