+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/UpdateHubRouteResponseDto.java
@@ -1,8 +1,12 @@
-package com.sparta.logistics.hub_service.dto;
+package com.sparta.logistics.hub_service.application.dto;
 
 import java.util.UUID;
+import lombok.Builder;
+import lombok.Getter;
 
-public class DeleteHubRouteResponseDto {
+@Getter
+@Builder
+public class UpdateHubRouteResponseDto {
   private UUID startHubId;
   private UUID endHubId;
   private String startHubName;
